+++ bundle/src/test/java/com/adobe/acs/commons/redirectmaps/impl/TestServlets.java
@@ -41,7 +41,7 @@ import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.resource.ModifiableValueMap;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.commons.testing.sling.MockResourceResolver;
-import org.apache.tika.io.IOUtils;
+import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;