+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ## Unreleased ([details][unreleased changes details])
 
+- #3715 - Add fallback logic for Page Root detection (Shared Component Properties) on Experience Fragments, Launches, and Version History
+
+### Fixed
+
+- #3742 Remove usage of Apache Tika 1.x
 
 ## 6.16.0 - 2026-02-10