+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ## Unreleased ([details][unreleased changes details])
 
+
+- #3732 Redirect Manager: evaluate "contextPrefixIgnored" for source path matching in addition to target path construction
 - #3715 - Add fallback logic for Page Root detection (Shared Component Properties) on Experience Fragments, Launches, and Version History
 
 ### Fixed