+++ CHANGELOG.md
@@ -8,15 +8,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->-
 
-## Unreleased ([details][unreleased changes details])
+## Unreleased
 
 - #3731 Redirect Manager: allow to delete all redirect manager rules below a common CAC path including the configuration node 
 
+## [6.17.0] - 2026-04-29
+
+## Changed
+
+- #3644 All Result checkbox on Reports could be configurable 
 - #3732 Redirect Manager: evaluate "contextPrefixIgnored" for source path matching in addition to target path construction
-- #3715 - Add fallback logic for Page Root detection (Shared Component Properties) on Experience Fragments, Launches, and Version History
+- #3734 RedirectManager: Support regular expression in source without capturing groups 
+- #3715 Add fallback logic for Page Root detection (Shared Component Properties) on Experience Fragments, Launches, and Version History
 
-### Fixed
+## Fixed
 
+- #3730 RedirectManager: Trailing "*" in source path being converted to "(.*)" might destroy regular expression
+- #3718 Redirect Manager: Fix replacing sharded redirects from a spreadsheet 
 - #3742 Remove usage of Apache Tika 1.x
 
 ## 6.16.0 - 2026-02-10